pyproject.toml
[project]
name = "devteam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.1",
    "google-adk>=1.24.1",
    "litellm>=1.81.9",
    "requests>=2.32.4",
    "streamlit>=1.54.0",
    "uvicorn[standard]>=0.34.0",
]
